Dan-Mark-SceneManager
We're Almost finished, All buttons have functionality and all work correctly with exception of the play again buutton, for this we need some sort of ResetGame() function in the GameManager. The score changes depending on the GameManager score, the Lives images change depending on the GameManager Lives, all seems to work as expeccted.