Mark-WinScreen.
I changed the gameOverMenu to endGameMenu, and added two text boxes which display the appropriate text at the and of the game, if all enemys are dead the win text displays, if all lives are lost the game over text displays. The win situation is stored in a function and called from the gameManager, this hasn't been tested yet.
I also added a line to the start button to call reset() from the gameManager, because if the player lost or won, went to the main menu, pressed the start button, the game didn't reset, it just displayed the previous game. If you want to do this a different way let me know, I mainly added it to test what I had done.