Added a highscore Manager
I created a high score manager from a singlton, this way the high score doesn't reset, when the game is started from the title screen. I moved all the high score code from the game manager, the canvas manager now calls the getHighScore()method from HSM instead of GM. everything works, I'm not sure if this was the best way to approach this so let me know if you have a different idea.