Music + Rearrangement of Prefabs + Re-editing of scores
SoundManager script created to manage the background music.
Sound also added to the death of the Light and Medium enemies as well as scores edited once again.

Currently experiencing issue when testing where enemies disappear, should not be occuring based on script and requires testing.

SoundManager used to switch audio playing based on the current Scene.
Still need to find how to add another audio file to the component other than the array, in order to utilise the starting sound for when the Game scene loads.

(May or may not have the wrong starting music playing :^) )